Move group index calculation setting to ModeSpec
