+++ tests/test_plugins/test_mode_solver.py
@@ -222,9 +222,8 @@ def test_group_index():
     ms = ModeSolver(
         simulation=simulation,
         plane=td.Box(size=(td.inf, td.inf, 0)),
-        mode_spec=mode_spec,
+        mode_spec=mode_spec.copy(update={"group_index_step": True}),
         freqs=[td.C_0 / 1.54, td.C_0 / 1.55, td.C_0 / 1.56],
-        compute_group_index=True,
     )
     modes = ms.solve()
     assert (modes.n_group.sel(mode_index=0).values > 3.9).all()
